@@ -35,15 +35,16 @@
   },
   "dependencies": {
     "history": "^3.0.0",
-    "react": "^15.3.2",
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4",
     "react-css-transition": "^0.7.4",
-    "react-dom": "^15.3.2",
+    "react-dom": "^15.5.4",
     "react-fastclick": "^3.0.1",
-    "react-ga": "^2.1.2",
+    "react-ga": "^2.2.0",
     "react-helmet": "^5.0.2",
-    "react-lite": "^0.15.33",
+    "react-lite": "^0.15.34",
     "react-router": "^3.0.0",
-    "react-router-scroll": "^0.4.1",
+    "react-router-scroll": "^0.4.2",
     "serve": "^5.1.2"
   },
   "devDependencies": {
@@ -53,9 +54,9 @@
     "babel-eslint": "^7.2.1",
     "babel-loader": "^6.4.1",
     "babel-preset-es2015": "^6.24.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-0": "^6.16.0",
-    "compression-webpack-plugin": "^0.3.2",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "compression-webpack-plugin": "^0.4.0",
     "copy-webpack-plugin": "^4.0.1",
     "cross-env": "^4.0.0",
     "css-loader": "^0.28.0",
@@ -85,10 +86,10 @@
     "postcss-custom-selectors": "^3.0.0",
     "postcss-import": "^9.1.0",
     "postcss-loader": "^1.2.2",
-    "postcss-nested": "^1.0.0",
+    "postcss-nested": "^1.0.1",
     "postcss-short": "^3.0.3",
-    "react-addons-test-utils": "^15.4.2",
-    "redbox-react": "^1.3.3",
+    "react-addons-test-utils": "^15.5.1",
+    "redbox-react": "^1.3.6",
     "rimraf": "^2.5.4",
     "sitemap": "^1.12.0",
     "style-loader": "^0.16.1",
