@@ -1,5 +1,7 @@
 /* eslint no-undef: 0 */
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+
 import { Link } from 'react-router';
 
 import Fader from '../../components/fader/';
