@@ -1,8 +1,25 @@
-import React from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 
 import styles from './style.css';
 
-class Radio extends React.Component {
+export default class Radio extends Component {
+
+  static propTypes = {
+    id: PropTypes.string,
+    name: PropTypes.string,
+    value: PropTypes.string,
+    label: PropTypes.string,
+    checked: PropTypes.string,
+  };
+
+  static defaultProps = {
+    id: '',
+    name: '',
+    value: '',
+    label: '',
+    checked: '',
+  };
 
   constructor(props) {
     super(props);
@@ -31,21 +48,3 @@ class Radio extends React.Component {
     );
   }
 }
-
-Radio.propTypes = {
-  id: React.PropTypes.string,
-  name: React.PropTypes.string,
-  value: React.PropTypes.string,
-  label: React.PropTypes.string,
-  checked: React.PropTypes.string,
-};
-
-Radio.defaultProps = {
-  id: '',
-  name: '',
-  value: '',
-  label: '',
-  checked: '',
-};
-
-export default Radio;
